+++ src/stories/components/Button.stories.ts
@@ -1,4 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
+import { fn } from '@storybook/test';
 import { Button } from '../../components';
 
 const meta = {
@@ -19,6 +20,7 @@ type Story = StoryObj<typeof meta>;
 export const Default: Story = {
   args: {
     text: 'Button',
+    onClick: fn(),
   },
 };
 
@@ -26,6 +28,7 @@ export const Colored: Story = {
   args: {
     text: 'Button',
     color: 'red',
+    onClick: fn(),
   },
 };
 
@@ -33,6 +36,7 @@ export const Disabeld: Story = {
   args: {
     text: 'Button',
     disabled: true,
+    onClick: fn(),
   },
 };
 
@@ -40,5 +44,6 @@ export const Half: Story = {
   args: {
     text: 'Button',
     half: true,
+    onClick: fn(),
   },
 };