+++ src/stories/components/Draw.stories.tsx
@@ -1,4 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
+import { fn } from '@storybook/test';
 import { Draw } from '../../components';
 
 const meta = {
@@ -7,6 +8,12 @@ const meta = {
   tags: ['autodocs'],
   argTypes: {
     //    backgroundColor: { control: 'color' },
+    width: {
+      control: 'number',
+    },
+    height: {
+      control: 'number',
+    },
   },
   parameters: {
     layout: 'centered',
@@ -14,11 +21,13 @@ const meta = {
 } satisfies Meta<typeof Draw>;
 
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
     width: 700,
     height: 500,
+    onChange: fn(),
   },
 };